src/main/webapp/app/entities/mefpai/route/mefpai-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IMefpai, Mefpai } from '../mefpai.model';
import { MefpaiService } from '../service/mefpai.service';

import { MefpaiRoutingResolveService } from './mefpai-routing-resolve.service';

describe('Mefpai routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: MefpaiRoutingResolveService;
  let service: MefpaiService;
  let resultMefpai: IMefpai | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(MefpaiRoutingResolveService);
    service = TestBed.inject(MefpaiService);
    resultMefpai = undefined;
  });

  describe('resolve', () => {
    it('should return IMefpai returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMefpai = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMefpai).toEqual({ id: 123 });
    });

    it('should return new IMefpai if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMefpai = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultMefpai).toEqual(new Mefpai());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Mefpai })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMefpai = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMefpai).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
